Move url parameter generation to a separate file
