+++ connectors/uplynk/src/main/java/com/theoplayer/android/connector/uplynk/internal/UplynkSsaiDescriptionConverter.kt
@@ -1,6 +1,5 @@
 package com.theoplayer.android.connector.uplynk.internal
 
-import com.theoplayer.android.connector.uplynk.UplynkAssetType
 import com.theoplayer.android.connector.uplynk.UplynkSsaiDescription
 import kotlin.time.Duration
 
@@ -18,45 +17,6 @@ internal class UplynkSsaiDescriptionConverter {
 
         return "$prefix/preplay/$urlAssetType/$urlAssetId?v=2$drmParameters$pingParameters$urlParameters"
     }
-
-    private val UplynkSsaiDescription.drmParameters: String
-        get() = if (contentProtected) {
-            "&manifest=mpd&rmt=wv"
-        } else {
-            ""
-        }
-
-    private val UplynkSsaiDescription.urlParameters
-        get() = if (preplayParameters.isNotEmpty()) {
-            preplayParameters.map { "${it.key}=${it.value}" }.joinToString("&", prefix = "&")
-        } else {
-            ""
-        }
-
-    private val UplynkSsaiDescription.pingParameters: String
-        get() {
-            val feature = UplynkPingFeatures.from(this)
-            return if (feature == UplynkPingFeatures.NO_PING) {
-                "&ad.pingc=0"
-            } else {
-                "&ad.pingc=1&ad.pingf=${feature.pingfValue}"
-            }
-        }
-
-    private val UplynkSsaiDescription.urlAssetType
-        get() = when (assetType) {
-            UplynkAssetType.ASSET -> ""
-            UplynkAssetType.CHANNEL -> "channel"
-            UplynkAssetType.EVENT -> "event"
-        }
-
-    private val UplynkSsaiDescription.urlAssetId
-        get() = when {
-            assetIds.isEmpty() && externalIds.size == 1 -> "$userId/${externalIds.first()}.json"
-            assetIds.isEmpty() && externalIds.size > 1 -> "$userId/${externalIds.joinToString(",")}/multiple.json"
-            assetIds.size == 1 -> "${assetIds.first()}.json"
-            else -> assetIds.joinToString(separator = ",") + "/multiple.json"
-        }
 
     fun buildAssetInfoUrls(
         ssaiDescription: UplynkSsaiDescription,
